Setup babel
