+++ package.json
@@ -17,6 +17,8 @@
   },
   "devDependencies": {
     "@babel/core": "^7.0.0-beta.40",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0-beta.40",
+    "@babel/preset-env": "^7.0.0-beta.40",
     "@types/clean-webpack-plugin": "^0.1.1",
     "@types/extract-text-webpack-plugin": "^3.0.1",
     "@types/glob": "^5.0.35",
